:fire: (CoreImu): Remove MLC components in CoreIMU
