--- drivers/CoreIMUlsm6dsox/include/interface/Accelerometer.h
@@ -1,38 +1,0 @@
-// Leka - LekaOS
-// Copyright 2021 APF France handicap
-// SPDX-License-Identifier: Apache-2.0
-
-
-#pragma once
-
-#include "LogKit.h"
-
-
-namespace leka::interface {
-
-    using mg_t = float;
-
-    class Accelerometer
-    {
-        public:
-            
-            struct Data {
-                mg_t x;
-                mg_t y;
-                mg_t z;
-            };
-            using accel_data_t = Data;
-
-            virtual ~Accelerometer() = default;
-
-            virtual auto getAccelX()                            -> mg_t     = 0;
-            virtual auto getAccelY()                            -> mg_t     = 0;
-            virtual auto getAccelZ()                            -> mg_t     = 0;
-            virtual auto getAccelXYZ() 			                -> std::tuple<mg_t, mg_t, mg_t>    = 0;
-            
-            virtual auto getAccelRate() 				        -> int32_t  = 0;
-            virtual auto getAccelRange() 				        -> int32_t  = 0;
-
-            virtual void turnOffAccel()                                     = 0;
-    };
-}
